// projects/angular/forms/datepicker/interfaces/day-of-week.interface.ts
/*
 * Copyright (c) 2016-2026 Broadcom. All Rights Reserved.
 * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
 * This software is released under MIT license.
 * The full license information can be found in LICENSE in the root directory of this project.
 */

export interface ClrDayOfWeek {
  readonly day: string;
  readonly narrow: string;
}

/**
 * Index of the day of the week, matching JavaScript's Date.getDay() convention.
 * Used to override the locale-derived first day of the week in the date picker.
 */
export enum ClrWeekday {
  Sunday = 0,
  Monday = 1,
  Tuesday = 2,
  Wednesday = 3,
  Thursday = 4,
  Friday = 5,
  Saturday = 6,
}